This is a work-in-progress stupid idea I had to put all libzfs related types and functions logic in ZFSWorker class, and then expose static functions in ZFSWorker to warp libzfs. My intention was to wrap libzfs in way that makes it easier to update the addon in case libzfs changes.
I got two worker to use the wrapper functions, but things started
getting unecessarily complicated. Also, the types exported by libzfs are
exported as they are via typedef. Maybe this is thing I can fix, but
overall, this is going to get very complex and hard to maintain in the
future. I will push this in seperate branch.

Who knows, maybe I will end up reviving this at some point in the
future.